refactor(app-framework): extract bootLoader driver into App/loader
Move the window.__bootLoader driver + its BootLoader/Window global types into app-framework's App/loader.ts (exported from the ui surface) so the React side and host apps can drive the boot loader without re-declaring the types; simplify App.tsx and composer-app main.tsx to consume it.
