+++ packages/sdk/app-framework/src/ui/components/App/index.ts
@@ -3,3 +3,4 @@
 //
 
 export * from './App';
+export * from './loader';
